FIX move E2eTestAccount class to own file
